Initial scaffolding and layout
Set up the modules of the flake. At this time, the flake is split into
modules and machines. Considering that additional top-level sections --
such as packages and shells -- will also be added, it might make sense
to keep all nix related things in one top-level directory, alongside the
flake.nix file.

This is all under the assumption that later amendments to the project
will introduce other tech, such as terraform.