Finalize working configs and prior to adding secrets
System configurations in a known good state. Secrets still have to be
integrated, at this point the nixos-anywhere project will likely handle
this to keep the bootstrap completely automatic since it can copy ssh
host keys to the host prior to install/activation. This would work
nicely with the sops secret module, but a good way to manage host keys
for each machine still has to be decided.