Push configuration out of flake
Projects are not necessarily all exported by the system flake, and
should freely support other derivations. The new "mod" directory will be
used for most projects, and it can be imported by machine
configurations.

That said, the tooling around flakes is pretty good and will still be
used until the dependency is lessened.